Capture all exceptions raised during extraction
This reorganizes the code in `aqt.installer.installer()` to catch all
exceptions derived from `Exception` (this excludes SystemExit,
KeyboardInterrupt, GeneratorExit) that can be raised during the archive
extraction step. These exceptions will be wrapped in an
ArchiveExtractionError.
The ArchiveExtractionError will provide suggestions to:
 a.) try a different extraction tool using `--external`, and
 b.) file a bug report if the error persists

WHY:
The point of this is to prevent the "PLEASE FILE A BUG REPORT" message
for bugs related to extraction. These bugs do not originate in aqt and
don't really belong in the aqt Issues section. Often, they can be
easily mitigated by using an external extractor.