Replace msvc2017 builds with mingw
MSVC 2017 is not available on the Windows-2019 host, so these builds must be
replaced with architectures that are available.

The existing mingw builds were a little bit broken because the `Path`
environment variable did not include the installed mingw toolchain, so
the Chocolatey mingw was being used. This change fixes that.