Print stack trace to DEBUG log
This change prints the stack trace for every AqtException to the
DEBUG log. This will make it easier to debug errors, in the case that
an exception is not being handled properly.
We will be able to ask our users to send us the log file when they
are reporting bugs. No matter what kind of exception occurred, there
will now be a record of it.