Improve tests for `helper.get_hash()`
This requires that `get_hash` does a better job of checking what data is
sent to it. To pass these tests, `get_hash` must check that the hash
is the right length, that the hash can be unhexlified properly, and
that it does not compose an unusable url when it tries to run `getUrl`.